+++ Hotel_API/Data/IAuthRepository.cs
@@ -6,7 +6,7 @@ namespace Hotel_API.Data
 
     public interface IAuthRepository
     {
-        public Task<ServiceResponse<int>> Register(User user, string password);
+        public Task<ServiceResponse<decimal>> Register(User user, string password);
         public Task<ServiceResponse<string>> Login(string email, string password);
         public Task<bool> UserExists(string email);
     }
